:recycle: Refactor to avoid circular imports in OrganizationNode and UserNode
