+++ creator/users/schema.py
@@ -1,25 +1,18 @@
 import graphene
-import django_filters
 from graphene import relay
 from graphql_relay import from_global_id
 from graphql import GraphQLError
 from graphene_django import DjangoObjectType
 from graphene_django.converter import convert_django_field_with_choices
-from graphene_django.filter import (
-    DjangoFilterConnectionField,
-    GlobalIDFilter,
-)
-from django_filters import OrderingFilter
+from graphene_django.filter import DjangoFilterConnectionField
 
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group, Permission
 
 from creator.studies.models import Membership
-from creator.users.mutations import Mutation
-
-from creator.organizations.nodes import OrganizationNode
 from creator.organizations.queries import OrganizationFilter
-
+from creator.users.mutations import Mutation
+from creator.users.queries import Query
 
 User = get_user_model()
 
@@ -100,7 +93,7 @@ class UserNode(DjangoObjectType):
     display_name = graphene.String(source="display_name")
 
     organizations = DjangoFilterConnectionField(
-        OrganizationNode,
+        "creator.organizations.queries.OrganizationNode",
         filterset_class=OrganizationFilter,
         description="List all organizations the user is in",
     )
@@ -189,104 +182,3 @@ class PermissionNode(DjangoObjectType):
             raise GraphQLError("Permission not found")
 
         return obj
-
-
-class UserFilter(django_filters.FilterSet):
-    organization = GlobalIDFilter(field_name="organizations")
-
-    class Meta:
-        model = User
-        fields = {
-            "email": ["exact", "contains"],
-            "username": ["exact"],
-            "first_name": ["exact"],
-            "last_name": ["exact"],
-            "last_login": ["gt", "lt"],
-            "date_joined": ["gt", "lt"],
-        }
-
-    order_by = OrderingFilter(fields=("date_joined",))
-
-
-class GroupFilter(django_filters.FilterSet):
-    name_contains = django_filters.CharFilter(
-        field_name="name", lookup_expr="contains"
-    )
-
-    class Meta:
-        model = Group
-        fields = {"name": ["exact"]}
-
-
-class PermissionFilter(django_filters.FilterSet):
-    name_contains = django_filters.CharFilter(
-        field_name="name", lookup_expr="contains"
-    )
-
-    class Meta:
-        model = Permission
-        fields = {"name": ["exact"], "codename": ["exact"]}
-
-
-class Query(object):
-    all_users = DjangoFilterConnectionField(
-        UserNode, filterset_class=UserFilter
-    )
-    my_profile = graphene.Field(UserNode)
-    group = relay.Node.Field(GroupNode, description="Get a group")
-    all_groups = DjangoFilterConnectionField(
-        GroupNode, filterset_class=GroupFilter, description="List all groups"
-    )
-    permission = relay.Node.Field(
-        PermissionNode, description="Get a permission"
-    )
-    all_permissions = DjangoFilterConnectionField(
-        PermissionNode,
-        filterset_class=PermissionFilter,
-        description="List all permissions",
-    )
-
-    def resolve_all_users(self, info, **kwargs):
-        """
-        If user is USER, only return that user
-        If user is ADMIN, return all users
-        If user is unauthed, return no users
-        """
-        user = info.context.user
-        if user is None or not user.is_authenticated:
-            return User.objects.none()
-
-        # Only return the current user if there are insufficient permissions
-        if user.is_authenticated and not user.has_perm(
-            "creator.list_all_user"
-        ):
-            return User.objects.filter(sub=user.sub).all()
-
-        return User.objects.all()
-
-    def resolve_my_profile(self, info, **kwargs):
-        """
-        Return the user that is making the request if they are valid,
-        otherwise, return nothing
-        """
-        user = info.context.user
-
-        # Unauthed and service users do not have profiles
-        if not user.is_authenticated or user is None or user.username is None:
-            raise GraphQLError("not authenticated as a user with a profile")
-
-        return user
-
-    def resolve_all_groups(self, info, **kwargs):
-        user = info.context.user
-        if not user.has_perm("auth.view_group"):
-            raise GraphQLError("Not allowed")
-
-        return Group.objects.all()
-
-    def resolve_all_permissions(self, info, **kwargs):
-        user = info.context.user
-        if not user.has_perm("auth.view_permission"):
-            raise GraphQLError("Not allowed")
-
-        return Permission.objects.all()
